Unit Testing for SummaryOffices
